packages/meeseeks_core/pyproject.toml
[project]
name = "meeseeks-core"
version = "2.0.0-alpha"
description = "Core module for Meeseeks - orchestration, schemas, and utilities."
readme = "../../README.md"
requires-python = ">=3.10,<4.0"
authors = [
  { name = "Krishnakanth Alagiri", email = "mail@kanth.tech" },
]
license = { text = "MIT" }

dependencies = [
  "python-dotenv>=1.0.1,<2.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "langchain-core>=1.2.8,<2.0.0",
  "langchain-community>=0.4.1,<0.5.0",
  "langchain-litellm>=0.4.0,<0.5.0",
  "litellm>=1.81.0,<2.0.0",
  "loguru>=0.7.2,<1.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "tiktoken>=0.7.0,<1.0.0",
  "typing-extensions>=4.12.2,<5.0.0",
]

[project.optional-dependencies]
observability = ["langfuse>=2.29.3,<3.0.0"]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meeseeks_core"]
include = ["src/meeseeks_core/prompts/**"]
